+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog for Nautilai
 ======================
 
-0.3.0 (unreleased)
+0.3.0 (2025-02-04)
 ------------------
 
 Added: